Add sentry & posthog logging (#21)
* Add Sentry

* Add sentry captures when certain failures loading drawing data happens

* Connect posthog to sentry
